requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-huggingface==0.0.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
google-generativeai==0.5.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
streamlit==1.33.0
ragas==0.1.9
datasets==2.19.1
plotly==5.21.0
Pillow==10.3.0
pypdf==4.2.0
huggingface-hub==0.23.0
tokenizers==0.19.1
transformers==4.40.2
